RavenDB-22142 Dedicated files for RefSchema and ElementSchemaRuleValidatorFactory. Dedicated ctor for ElementSchemaRuleValidator without types restrictions
